uibuild

+++ controllers/blogs.js
@@ -1,15 +1,23 @@
 const blogsRouter = require("express").Router();
-const { response } = require("../app");
+const jwt = require("jsonwebtoken");
 const Blog = require("../models/blog");
+const User = require("../models/user");
 
-// blogsRouter.get("/", (request, response) => {
+// const getTokenFrom = (request) => {
+//   const authorization = request.get("authorization");
+//   if (authorization && authorization.startsWith("Bearer ")) {
+//     return authorization.replace("Bearer ", "");
+//   }
+//   return null;
+// };
+// blogsRouter.get("/", (request, response) => {s
 //   Blog.find({}).then((blogs) => {
 //     response.json(blogs);
 //   });
 // });
 
 blogsRouter.get("/", async (request, response) => {
-  const blogs = await Blog.find({});
+  const blogs = await Blog.find({}).populate("user", { username: 1, name: 1 });
   response.json(blogs);
 });
 
@@ -25,19 +33,36 @@ blogsRouter.get("/", async (request, response) => {
 // });
 
 blogsRouter.delete("/:id", async (request, response) => {
+  const blog = await Blog.findByIdAndDelete(request.params.id);
+  const user = request.user;
+  const decodedToken = jwt.verify(request.token, process.env.SECRET);
+  if (blog.user.toString() !== decodedToken.id) {
+    return response
+      .status(403)
+      .json({ error: "Permission denied: You can only delete your own blogs" });
+  }
   await Blog.findByIdAndDelete(request.params.id);
   response.status(204).end();
 });
 
 blogsRouter.post("/", async (request, response) => {
   const body = request.body;
+  const decodedToken = jwt.verify(request.token, process.env.SECRET);
+  if (!decodedToken.id) {
+    return response.status(401).json({ error: "token invalid" });
+  }
+  const user = request.user;
+
   const blog = new Blog({
     title: body.title,
     author: body.author,
     url: body.url,
     likes: body.likes || 0,
+    user: user.id,
   });
   const savedBlog = await blog.save();
+  user.blogs = user.blogs.concat(savedBlog._id);
+  await user.save();
   response.status(201).json(savedBlog);
 });
